// src/types/Enums/HeroClasses.ts
export enum HeroClasses {
	KNIGHT = 'Рыцарь',
	CLERIC = 'Клерик',
	RANGER = 'Рейнджер',
	DRUID = 'Друид',
	ALCHEMIST = 'Алхимик',
	WIZARD = 'Волшебник',
	DEMONIAC = 'Одержимый',
	HERETIC = 'Еретик',
	DEATH_KNIGHT = 'Рыцарь Смерти',
	NECROMANCER = 'Некромант',
	OVERLORD = 'Верховный Лорд',
	WARLOCK = 'Чернокнижник',
	BARBARIAN = 'Варвар',
	BATTLE_MAGE = 'Боевой Маг',
	BEASTMASTER = 'Зверолов',
	WITCH = 'Ведьма',
	PLANESWALKER = 'Странник',
	ELEMENTALIST = 'Элементалист',
}