// src/types/Enums/Spells.ts
export enum Spells {
  STONE_SKIN = 'Каменная Кожа',
  BLESS = 'Благословение',
  FROST_RING = 'Морозное кольцо',
  CURE = 'Лечение',
  WEAKNESS = 'Слабость',
  CURSE = 'Проклятие',
  PRAYER = 'Молитва',
  DISPEL = 'Снятие Заклятий',
  DISRUPTING_RAY = 'Разрушающий Луч',
  HASTE = 'Ускорение',
  MAGIC_ARROW = 'Магическая Стрела',
  SLOW = 'Замедление',
  BLOOD_LUST = 'Кровожадность',
  FIRE_WALL = 'Огненная Стена',
  RESURRECTION = 'Воскрешение',
  METEOR_SHOWER = 'Метеоритный Дождь',
  SHIELD = 'Щит',
  PROTECTION_FROM_AIR = 'Защита От Воздуха',
  PROTECTION_FROM_FIRE = 'Защита От Огня',
  VIEW_EARTH = 'просмотр Земли',
  FIREBALL = 'Огненный шар',
  INFERNO = 'Инферно',
  DEATH_RIPPLE = 'Волна Смерти',
  ANIMATE_DEAD = 'Оживление Мертвецов',
  ICE_BOLT = 'Ледяная Стрела',
  SLAYER = 'Убийца',
  SUMMON_BOAT = 'Вызвать лодку',
  FORTUNE = 'Удача',
  PRECISION = 'Точность',
  HYPNOTIZE = 'Гипноз',
  CHAIN_LIGHTNING= 'Цепная Молния',
}